completed version
Write a static method named triathlon that accepts as its parameter a Scanner for an input file whose data represents triathlon race results for athletes. Your method should add up the swimming, biking, and running times for each athlete and report the total time for each athlete and their time difference from the winner's. The input consists of a series of tokens. Each athlete's data is represented by four tokens in the following order: athlete's first name, swimming time, biking time, and running time (all times are given in minutes). The data for an athlete may or may not span multiple lines, but you are guaranteed the athletes will come in the order they finished the race (the winner's data is first, the second place triathlete's data comes next, etc.).